Move stores to stores module
