+++ src/components/channel/index.tsx
@@ -4,8 +4,8 @@ import { useEffect } from "react";
 import { useState } from "react";
 import { Channel } from "../../models/channels";
 import Mixer from "../../models/mixer";
+import editModeStore from "../../stores/editModeStore";
 import EmptyChannel from "../emptyChannel";
-import editModeStore from "../mixerActions/store";
 import ChannelFader from "./addOns/channelFader";
 import ChannelName from "./addOns/channelName";
 import MuteSoloComponent from "./addOns/muteSolo";
